+++ packages/s2-core/__tests__/unit/facet/table-facet-spec.ts
@@ -71,7 +71,7 @@ describe('Table Mode Facet Test', () => {
   const dataSet: TableDataSet = new MockTableDataSet(ss);
   const facet: TableFacet = new TableFacet({
     spreadsheet: ss,
-    dataSet: dataSet,
+    dataSet,
     ...assembleDataCfg().fields,
     ...assembleOptions(),
     ...DEFAULT_STYLE,
@@ -93,7 +93,7 @@ describe('Table Mode Facet Test With Adaptive Layout', () => {
   const dataSet: TableDataSet = new MockTableDataSet(ss);
   const options = {
     spreadsheet: ss,
-    dataSet: dataSet,
+    dataSet,
     ...assembleDataCfg().fields,
     ...assembleOptions({}),
     ...DEFAULT_STYLE,
@@ -174,7 +174,7 @@ describe('Table Mode Facet Test With Compact Layout', () => {
     };
     const facet: TableFacet = new TableFacet({
       spreadsheet: ss,
-      dataSet: dataSet,
+      dataSet,
       ...assembleDataCfg().fields,
       ...assembleOptions(),
       ...DEFAULT_STYLE,
@@ -185,7 +185,7 @@ describe('Table Mode Facet Test With Compact Layout', () => {
     test('col hierarchy coordinate with compact layout', () => {
       const { colLeafNodes } = facet.layoutResult;
 
-      const COMPACT_WIDTH = [52, 52, 64, 40, 72];
+      const COMPACT_WIDTH = [53, 53, 65, 41, 73];
 
       let lastX = 0;
       colLeafNodes.forEach((node, index) => {
@@ -206,7 +206,7 @@ describe('Table Mode Facet Test With Compact Layout', () => {
     };
     const facet: TableFacet = new TableFacet({
       spreadsheet: ss,
-      dataSet: dataSet,
+      dataSet,
       ...assembleDataCfg().fields,
       ...assembleOptions(),
       ...DEFAULT_STYLE,
@@ -218,7 +218,7 @@ describe('Table Mode Facet Test With Compact Layout', () => {
     test('col hierarchy coordinate with compact layout with seriesNumber', () => {
       const { colLeafNodes } = facet.layoutResult;
 
-      const COMPACT_WIDTH = [80, 52, 52, 64, 40, 72];
+      const COMPACT_WIDTH = [80, 53, 53, 65, 41, 73];
 
       let lastX = 0;
       colLeafNodes.forEach((node, index) => {
@@ -237,7 +237,7 @@ describe('Table Mode Facet With Frozen Test', () => {
   const dataSet: TableDataSet = new MockTableDataSet(ss);
   const facet: TableFacet = new TableFacet({
     spreadsheet: ss,
-    dataSet: dataSet,
+    dataSet,
     ...assembleDataCfg().fields,
     ...assembleOptions({
       frozenColCount: 2,
@@ -357,6 +357,20 @@ describe('Table Mode Facet With Frozen Test', () => {
     expect(viewCellHeights.getCellOffsetY(0)).toBe(0);
     expect(viewCellHeights.getCellOffsetY(7)).toBe(210);
   });
+
+  test('should get correct indexes with row height gt canvas height', () => {
+    const originHeight = facet.panelBBox.viewportHeight;
+    facet.panelBBox.viewportHeight = 10;
+    expect(facet.calculateXYIndexes(0, 0)).toStrictEqual({
+      center: [2, 2, 2, 0],
+      frozenCol: [0, 1, 2, 0],
+      frozenRow: [2, 2, 0, 1],
+      frozenTrailingCol: [3, 4, 2, 0],
+      frozenTrailingRow: [2, 2, 30, 31],
+    });
+    // reset
+    facet.panelBBox.viewportHeight = originHeight;
+  });
 });
 
 describe('Table Mode Facet Test With Custom Row Height', () => {
@@ -376,7 +390,7 @@ describe('Table Mode Facet Test With Custom Row Height', () => {
   });
   const facet: TableFacet = new TableFacet({
     spreadsheet: ss,
-    dataSet: dataSet,
+    dataSet,
     ...assembleDataCfg().fields,
     ...merge({}, assembleOptions()),
     ...DEFAULT_STYLE,
@@ -420,7 +434,7 @@ describe('Table Mode Facet Test With Zero Height', () => {
   });
   const facet: TableFacet = new TableFacet({
     spreadsheet: ss,
-    dataSet: dataSet,
+    dataSet,
     ...assembleDataCfg().fields,
     ...merge({}, assembleOptions()),
     ...DEFAULT_STYLE,
@@ -445,7 +459,7 @@ describe('Table Mode Facet With Frozen layoutCoordinate Test', () => {
   const dataSet: TableDataSet = new MockTableDataSet(ss);
   const facet: TableFacet = new TableFacet({
     spreadsheet: ss,
-    dataSet: dataSet,
+    dataSet,
     ...assembleDataCfg().fields,
     ...assembleOptions({
       frozenColCount: 2,