+++ packages/s2-core/__tests__/unit/interaction/brush-selection-spec.ts
@@ -252,7 +252,7 @@ describe('Interaction Brush Selection Tests', () => {
     ).toBeFalsy();
   });
 
-  test('should get brush selection range cells', async () => {
+  test('should get brush selection range cells', () => {
     const selectedFn = jest.fn();
     const brushSelectionFn = jest.fn();
 
@@ -317,7 +317,7 @@ describe('Interaction Brush Selection Tests', () => {
     expect(brushSelectionFn).toHaveBeenCalledTimes(1);
   });
 
-  test('should get correct formatted brush point', async () => {
+  test('should get correct formatted brush point', () => {
     const EXTRA_PIXEL = 2;
     const VSCROLLBAR_WIDTH = 5;
     const { width, height } = mockSpreadSheetInstance.facet.getCanvasHW();
@@ -409,7 +409,7 @@ describe('Interaction Brush Selection Tests', () => {
     });
   });
 
-  test('shoud get correct selected cell metas', async () => {
+  test('should get correct selected cell metas', () => {
     expect(
       brushSelectionInstance.getSelectedCellMetas({
         start: {
@@ -424,7 +424,7 @@ describe('Interaction Brush Selection Tests', () => {
     ).toBe(100);
   });
 
-  test('shoud get correct adjusted frozen rowIndex and colIndex', async () => {
+  test('should get correct adjusted frozen rowIndex and colIndex', () => {
     const { adjustNextColIndexWithFrozen, adjustNextRowIndexWithFrozen } =
       brushSelectionInstance;
     mockSpreadSheetInstance.setOptions({
@@ -451,7 +451,7 @@ describe('Interaction Brush Selection Tests', () => {
     expect(adjustNextRowIndexWithFrozen(7, ScrollDirection.TRAILING)).toBe(7);
   });
 
-  test('shoud get correct scroll offset for row and col', async () => {
+  test('should get correct scroll offset for row and col', () => {
     const { facet } = mockSpreadSheetInstance;
     expect(
       getScrollOffsetForCol(
@@ -537,7 +537,7 @@ describe('Interaction Brush Selection Tests', () => {
     ).toBe(420);
   });
 
-  test('shoud get valid x and y index', async () => {
+  test('should get valid x and y index', () => {
     const { validateXIndex, validateYIndex } = brushSelectionInstance;
     expect(validateXIndex(-1)).toBe(null);
     expect(validateXIndex(1)).toBe(1);